trading_app/requirements.txt
asgiref==3.8.1
Django==5.1
django-crontab==0.7.1
djangorestframework==3.15.2
drf-spectacular==0.27.2
inflection==0.5.1
sqlparse==0.5.1
uritemplate==4.1.1
